pyproject.toml
[project]
name = "vcer"
version = "0.1.0"
description = "Visual Context Engineering Router: analyze, visualize, optimize, and route prompts to local LLM backends (vLLM/sglang)."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Your Name" }
]
license = { text = "Proprietary" }
keywords = ["LLM", "prompt", "context-engineering", "vllm", "sglang", "router"]

dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pydantic>=2.8",
  "ruamel.yaml>=0.18",
  "httpx>=0.27",
  "jinja2>=3.1",
  "markdown-it-py>=3.0",
]

[project.scripts]
vcer = "vcer.__main__:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
# uv will use this metadata to manage the project environment.
dev-dependencies = [
  "pytest>=8.3",
  "ruff>=0.6",
]

